fix mer merge conflicts
